Behaviour package restructure.
