+++ src/main/java/lasersharks/behaviour/checkforloss/DefaultCheckForLossBehaviour.java
@@ -1,6 +1,6 @@
-package lasersharks.behaviour.collision;
+package lasersharks.behaviour.checkforloss;
 
-import lasersharks.behaviour.collision.interfaces.CheckForLossBehaviour;
+import lasersharks.behaviour.interfaces.CheckForLossBehaviour;
 
 /**
  * Check for loss behaviour for SeaObjects.
