+++ src/main/java/lasersharks/behaviour/eaten/AmmoEatenBehaviour.java
@@ -1,7 +1,7 @@
-package lasersharks.behaviour.collision;
+package lasersharks.behaviour.eaten;
 
 import lasersharks.Logger;
-import lasersharks.behaviour.collision.interfaces.EeatenBehaviour;
+import lasersharks.behaviour.interfaces.EeatenBehaviour;
 import lasersharks.seaobjects.SeaObject;
 
 /**
