+++ src/main/java/lasersharks/behaviour/eaten/DefaultEatenBehaviour.java
@@ -1,6 +1,6 @@
-package lasersharks.behaviour.collision;
+package lasersharks.behaviour.eaten;
 
-import lasersharks.behaviour.collision.interfaces.EeatenBehaviour;
+import lasersharks.behaviour.interfaces.EeatenBehaviour;
 
 /**
  * Eaten behaviour for SeaObjects.
