+++ src/main/java/lasersharks/behaviour/interfaces/AmmunitionIncrementBehaviour.java
@@ -1,4 +1,4 @@
-package lasersharks.behaviour.collision.interfaces;
+package lasersharks.behaviour.interfaces;
 
 public interface AmmunitionIncrementBehaviour {
   int onCollisionAmmunitionIncrement();
