+++ src/main/java/lasersharks/behaviour/interfaces/CheckForLossBehaviour.java
@@ -1,4 +1,4 @@
-package lasersharks.behaviour.collision.interfaces;
+package lasersharks.behaviour.interfaces;
 
 public interface CheckForLossBehaviour {
   void onCollisionPlayerLoses(int size);
