+++ src/main/java/lasersharks/behaviour/interfaces/EeatenBehaviour.java
@@ -1,4 +1,4 @@
-package lasersharks.behaviour.collision.interfaces;
+package lasersharks.behaviour.interfaces;
 
 public interface EeatenBehaviour {
   void onCollisionEaten();
