+++ src/main/java/lasersharks/behaviour/interfaces/LaserCollisionBehaviour.java
@@ -1,4 +1,4 @@
-package lasersharks.behaviour.collision.interfaces;
+package lasersharks.behaviour.interfaces;
 
 public interface LaserCollisionBehaviour {
   boolean onCollisionDestroyLaser();
