+++ src/main/java/lasersharks/behaviour/interfaces/SizeDecrementBahaviour.java
@@ -1,4 +1,4 @@
-package lasersharks.behaviour.collision.interfaces;
+package lasersharks.behaviour.interfaces;
 
 public interface SizeDecrementBahaviour {
   void onCollisionSizeDecrement(int size);
