+++ src/main/java/lasersharks/behaviour/sizedecrement/FishSizeDecrementBehaviour.java
@@ -1,6 +1,6 @@
-package lasersharks.behaviour.collision;
+package lasersharks.behaviour.sizedecrement;
 
-import lasersharks.behaviour.collision.interfaces.SizeDecrementBahaviour;
+import lasersharks.behaviour.interfaces.SizeDecrementBahaviour;
 import lasersharks.seaobjects.SeaObject;
 
 /**
